pyproject.toml
[tool.poetry]
name = "ImageTextEraser"
version = "0.1.0"
description = ""
authors = ["MLRichter <matrichter@uos.de>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.5.5"
scikit-image = "^0.19.2"
imutils = "^0.5.4"
attrs = "^21.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
